Updated GN binaries, tweaked build scripts, added bgfx initialization failsafe to use OpenGL as a backup

+++ engine/include/purpl/core/inst.h
@@ -35,10 +35,10 @@ struct purpl_instance {
 	time_t start_time; // When the engine started
 	
 	SDL_Window *wnd; // The main window for the engine
-	s16 wnd_width; // The width of the window
-	u16 wnd_height; // The height of the window
-	u16 wnd_x; // The X position of the window
-	u16 wnd_y; // The Y position of the window
+	u32 wnd_width; // The width of the window
+	u32 wnd_height; // The height of the window
+	u32 wnd_x; // The X position of the window
+	u32 wnd_y; // The Y position of the window
 };
 
 extern struct purpl_instance *purpl_inst;
